let circleDiameter = (guide.extraJson?["circle_diameter"] as? CGFloat) ?? 200

+++ Sources/GuideBlocks/CircleVideo/CircleVideoGuide.swift
@@ -44,7 +44,7 @@ public class CircleVideoGuide: CTXBaseGuideController {
         let guide = contextualContainer.guidePayload.guide
         let defaultVideoURL = "https://www.youtube.com/embed/Y9ChGCY8Azk?si=aLGas88lnxI6g_jJ?autoplay=1"
         let videoUrl = (guide.extraJson?["vid_url"] as? String) ?? defaultVideoURL
-        let circleDiameter = (guide.extraJson?["circle_diameter"] as? Int) ?? 200
+        let circleDiameter = (guide.extraJson?["circle_diameter"] as? CGFloat) ?? 200
         
         let view = CircleVideoView(
             imageElement: guide.arrayImages.first,
